Fix broken migration w/drop index not existing yet

+++ db/migrate/20200821130027_add_missing_unique_indices.acts_as_taggable_on_engine.rb
@@ -9,7 +9,7 @@ AddMissingUniqueIndices.class_eval do
     add_index ActsAsTaggableOn.tags_table, :name, unique: true
 
     remove_index ActsAsTaggableOn.taggings_table, :tag_id if index_exists?(ActsAsTaggableOn.taggings_table, :tag_id)
-    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx'
+    remove_index ActsAsTaggableOn.taggings_table, name: 'taggings_taggable_context_idx' if index_exists?(ActsAsTaggableOn.taggings_table, :name)
     add_index ActsAsTaggableOn.taggings_table,
               [:tag_id, :taggable_id, :taggable_type, :context, :tagger_id, :tagger_type],
               unique: true, name: 'taggings_idx'
